Add-Migration for Order info

+++ MyProject/Entities/OrderDetail.cs
@@ -8,7 +8,7 @@ namespace Entities
     {
         [Key]
         public int OrderDetailId { get; set; }
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
         public double Price { get; set; }
